curso/services/sessions/pyproject.toml
[project]
name = "sessions"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.6",
    "opentelemetry-api>=1.30.0",
    "opentelemetry-exporter-otlp>=1.30.0",
    "opentelemetry-exporter-prometheus>=0.51b0",
    "opentelemetry-instrumentation-fastapi>=0.51b0",
    "opentelemetry-instrumentation-redis>=0.51b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.51b0",
    "opentelemetry-sdk>=1.30.0",
    "opentelemetry-semantic-conventions>=0.51b0",
    "prometheus-client>=0.21.1",
    "psycopg2-binary>=2.9.10",
    "py-dateutil>=2.2",
    "pydantic>=2.10.5",
    "pyjwt>=2.10.1",
    "pyyaml>=6.0.2",
    "redis>=5.2.1",
    "sqlalchemy>=2.0.37",
    "structlog>=25.1.0",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "alembic>=1.14.0",
]
